Rename UnitCompDTO to UnitCompDTOTemplate
